app/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
fake-headers==1.0.2
html5lib==1.1
idna==3.4
lxml==4.9.2
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
